Split out just ability score
